+++ tests/test_base_repository.py
@@ -25,37 +25,39 @@ def test_get_all(app_ctx, database_session):
     tyre_model_1 = repo.create(manufacturer='Michelin', model_name='Test Tyre Model')
     tyre_model_2 = repo.create(manufacturer='Not Michelin', model_name='Different Model Name')
     tyre_model_3 = repo.create(manufacturer='Pirelli', model_name='Third Model')
+    tyre_model_4 = repo.create(manufacturer='New Manufacturer', model_name='Fourth Model')
 
     # Can get all results with no pagination
     results, total_count = repo.get_all()
 
-    assert len(results) == 3
-    assert total_count == 3
+    assert len(results) == 4
+    assert total_count == 4
     assert results[0].id == tyre_model_1.id
     assert results[1].id == tyre_model_2.id
     assert results[2].id == tyre_model_3.id
+    assert results[3].id == tyre_model_4.id
 
     # Can limit number of results returned with pagination
-    limit_results, total_count = repo.get_all(limit=2)
+    limit_results, total_count = repo.get_all(page_size=2)
 
     assert len(limit_results) == 2
-    assert total_count == 3
+    assert total_count == 4
     assert limit_results[0].id == tyre_model_1.id
     assert limit_results[1].id == tyre_model_2.id
 
-    # Can offset start point of pagination
-    offset_results, total_count = repo.get_all(limit=2, offset=1)
+    # Can get page 2 of results
+    page_2_results, total_count = repo.get_all(page_size=2, page=2)
 
-    assert len(offset_results) == 2
-    assert total_count == 3
-    assert offset_results[0].id == tyre_model_2.id
-    assert offset_results[1].id == tyre_model_3.id
+    assert len(page_2_results) == 2
+    assert total_count == 4
+    assert page_2_results[0].id == tyre_model_3.id
+    assert page_2_results[1].id == tyre_model_4.id
 
     # Can not return results with invalid offset
-    empty_results, total_count = repo.get_all(offset=10)
+    empty_results, total_count = repo.get_all(page=10)
 
     assert len(empty_results) == 0
-    assert total_count == 3
+    assert total_count == 4
 
 def test_get_by_id(app_ctx, database_session):
     repo = BaseRepository(TyreModel)